Introduced option to show headline (off by default)
It works reasonably, but the frametitle will shift upward
correspondingly. It might be nice to figure out how to retain the
frametitle in the same vertical position.
It has been tried to simply insert an empty beamercolorbox with the dark
blue colour instead of the headline when the headline is off, but this
does not work, because it results in a tiny vertical gap between this
empty beamercolorbox and the frametitle background rectangle.