New option to disable headline improved and simplified
The frametitle is now in the same vertical position regardless of the
presence of the headline. It is now achieved by drawing an empty, dark
blue headline beamercolorbox instead of the white one used when the
headline is on. The previously observed gap above the frametitle
background rectangle is now simply covered up by shifting its upper left
corner slightly up (slightly dirty solution).